refactor(transport): move all remote stream related messages from protocol to transport
Both client and server for remote stream management will be implemented in transport module. So there is no need to define it in the protocol. Besides
