+++ protocol/src/main/resources/protocol.xml
@@ -86,10 +86,6 @@
       <type name="patchVersion" primitiveType="int32"/>
     </composite>
 
-    <composite name="UUID" description="Universally Unique ID">
-      <type name="high" primitiveType="int64" />
-      <type name="low" primitiveType="int64" />
-    </composite>
   </types>
 
   <!-- L1 General Messages 0 - 99 -->
@@ -169,21 +165,5 @@
       <field name="healthStatus" id="17" type="PartitionHealthStatus"/>
     </group>
     <data name="version" id="14" type="varDataEncoding"/>
-  </sbe:message>
-
-  <!-- L3 Gateway Stream messages 400-499 -->
-  <sbe:message name="AddStreamRequest" id="400" description="Adds a gateway stream to a broker">
-    <field name="id" id="1" type="UUID" />
-    <data name="streamType" id="2" type="varDataEncoding" />
-    <data name="metadata" id="3" type="varDataEncoding" />
-  </sbe:message>
-
-  <sbe:message name="RemoveStreamRequest" id="401" description="Removes a gateway stream from a broker">
-    <field name="id" id="1" type="UUID" />
-  </sbe:message>
-
-  <sbe:message name="PushStreamRequest" id="402" description="Pushes a payload over a stream">
-    <field name="id" id="1" type="UUID" />
-    <data name="payload" id="2" type="varDataEncoding"/>
   </sbe:message>
 </sbe:messageSchema>
