+++ transport/src/main/java/io/camunda/zeebe/transport/stream/impl/RemoteStreamApiServer.java
@@ -12,10 +12,10 @@ import io.atomix.cluster.ClusterMembershipEvent.Type;
 import io.atomix.cluster.ClusterMembershipEventListener;
 import io.atomix.cluster.MemberId;
 import io.atomix.cluster.messaging.ClusterCommunicationService;
-import io.camunda.zeebe.protocol.impl.stream.AddStreamRequest;
-import io.camunda.zeebe.protocol.impl.stream.JobStreamTopics;
-import io.camunda.zeebe.protocol.impl.stream.RemoveStreamRequest;
 import io.camunda.zeebe.scheduler.Actor;
+import io.camunda.zeebe.transport.stream.impl.messages.AddStreamRequest;
+import io.camunda.zeebe.transport.stream.impl.messages.JobStreamTopics;
+import io.camunda.zeebe.transport.stream.impl.messages.RemoveStreamRequest;
 import io.camunda.zeebe.util.buffer.BufferReader;
 import java.util.function.Function;
 import org.agrona.concurrent.UnsafeBuffer;
