+++ transport/src/main/java/io/camunda/zeebe/transport/stream/impl/messages/JobStreamTopics.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.protocol.impl.stream;
+package io.camunda.zeebe.transport.stream.impl.messages;
 
 public enum JobStreamTopics {
   ADD("job-stream-add"),
