+++ transport/src/main/java/io/camunda/zeebe/transport/stream/impl/messages/PushStreamRequest.java
@@ -5,12 +5,8 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.protocol.impl.stream;
+package io.camunda.zeebe.transport.stream.impl.messages;
 
-import io.camunda.zeebe.protocol.record.MessageHeaderDecoder;
-import io.camunda.zeebe.protocol.record.MessageHeaderEncoder;
-import io.camunda.zeebe.protocol.record.PushStreamRequestDecoder;
-import io.camunda.zeebe.protocol.record.PushStreamRequestEncoder;
 import io.camunda.zeebe.util.SbeUtil;
 import io.camunda.zeebe.util.buffer.BufferReader;
 import io.camunda.zeebe.util.buffer.BufferWriter;
