+++ transport/src/test/java/io/camunda/zeebe/transport/stream/impl/messages/SerializationTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.protocol.impl.stream;
+package io.camunda.zeebe.transport.stream.impl.messages;
 
 import static org.assertj.core.api.Assertions.assertThat;
